improve rules
